+++ resources/views/admin/pengaturan_surat/index.blade.php
@@ -17,20 +17,14 @@
     <div class="box box-info">
         <div class="box-header with-border">
             @if (can('u'))
-                <a href="{{ route('surat_master.form') }}" title="Tambah Format Surat"
-                    class="btn btn-social bg-olive btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i
-                        class="fa fa-plus"></i> Tambah</a>
+                <a href="{{ route('surat_master.form') }}" title="Tambah Format Surat" class="btn btn-social bg-olive btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i class="fa fa-plus"></i> Tambah</a>
             @endif
             @if (can('h'))
-                <a href="#confirm-delete" title="Hapus Data"
-                    onclick="deleteAllBox('mainform','{{ route('surat_master/deleteAll') }}')"
-                    class="btn btn-social btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"><i
+                <a href="#confirm-delete" title="Hapus Data" onclick="deleteAllBox('mainform','{{ route('surat_master/deleteAll') }}')" class="btn btn-social btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"><i
                         class='fa fa-trash-o'></i> Hapus</a>
             @endif
-            @if (! setting('nonaktifkan_rtf') && can('u', '', true))
-                <a href="{{ route('surat_master.perbarui') }}" title="{{ SebutanDesa('Perbarui Surat [Desa]') }}"
-                    class="btn btn-social bg-orange btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i
-                        class="fa fa-recycle"></i> Perbarui</a>
+            @if (!setting('nonaktifkan_rtf') && can('u', '', true))
+                <a href="{{ route('surat_master.perbarui') }}" title="{{ SebutanDesa('Perbarui Surat [Desa]') }}" class="btn btn-social bg-orange btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i class="fa fa-recycle"></i> Perbarui</a>
             @endif
             @if (can('u'))
                 <div class="btn-group-vertical radius-3">
@@ -38,20 +32,22 @@
                         Impor / Ekspor</a>
                     <ul class="dropdown-menu" role="menu">
                         <li>
-                            <a href="{{ route('surat_master.impor') }}" class="btn btn-social btn-block btn-sm"
-                                data-target="#impor-surat" data-remote="false" data-toggle="modal" data-backdrop="false"
-                                data-keyboard="false"><i class="fa fa-upload"></i> Impor Surat TinyMCE</a>
+                            <a
+                                href="{{ route('surat_master.impor') }}"
+                                class="btn btn-social btn-block btn-sm"
+                                data-target="#impor-surat"
+                                data-remote="false"
+                                data-toggle="modal"
+                                data-backdrop="false"
+                                data-keyboard="false"
+                            ><i class="fa fa-upload"></i> Impor Surat TinyMCE</a>
                         </li>
                         <li>
-                            <a target="_blank"
-                                class="btn btn-social btn-block btn-sm aksi-terpilih" title="Ekspor Surat TinyMCE" onclick="formAction('mainform', '{{ route('surat_master.ekspor') }}'); return false;"><i
-                                    class="fa fa-download"></i> Ekspor Surat TinyMCE</a>
+                            <a target="_blank" class="btn btn-social btn-block btn-sm aksi-terpilih" title="Ekspor Surat TinyMCE" onclick="formAction('mainform', '{{ route('surat_master.ekspor') }}'); return false;"><i class="fa fa-download"></i> Ekspor Surat TinyMCE</a>
                         </li>
                     </ul>
                 </div>
-                <a href="{{ route('surat_master.pengaturan') }}" title="Pengaturan"
-                    class="btn btn-social bg-purple btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i
-                        class="fa fa-gear"></i> Pengaturan</a>
+                <a href="{{ route('surat_master.pengaturan') }}" title="Pengaturan" class="btn btn-social bg-purple btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i class="fa fa-gear"></i> Pengaturan</a>
             @endif
 
             @if (ENVIRONMENT === 'development')
